+++ .travis.yml
@@ -19,8 +19,6 @@ env:
   - USER_LANGUAGE=en USER_REGION=US'
   - USER_LANGUAGE=fr USER_REGION=FR'
   - USER_LANGUAGE=ja USER_REGION=JP'
-  - USER_LANGUAGE=pt USER_REGION=BR'
-  - USER_LANGUAGE=default USER_REGION=default
 
 os:
   - linux
